% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea_api_key.R
\name{bea_key}
\alias{bea_key}
\title{Establishes a U.S. Bureau of Economic Analysis API key for \code{\link[=retrieve_data]{retrieve_data()}}}
\usage{
bea_key(key)
}
\arguments{
\item{key}{Supply your Bureau of Economic Analysis API key.}
}
\description{
If you haven't yet registered for a Bureau of Economic Analysis API key, follow the instructions at http://www.bea.gov/API/signup/index.cfm.
}
