% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories_vars.R
\name{categories_vars}
\alias{categories_vars}
\title{Returns variables from categories for \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}}
\usage{
categories_vars(categories)
}
\arguments{
\item{categories}{Supply a vector of categories you wish to return all
variables associated with, as listed in the \code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} data
frame.}
}
\description{
Takes a vector of one or more categories, as listed in the
\code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} data frame, and returns all variables associated
with those categories. Primarily intended for use inside the
\code{\link[=retrieve_aggregated]{retrieve_aggregated()}} function parentheses.
}
