% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_by_msa_2024}
\alias{crime_by_msa_2024}
\title{Crimes by Metropolitan Statistical Area in the United States in 2024}
\format{
A data frame with 3087 rows and 3 variables:
\describe{
\item{msa}{metropolitan statistical area}
\item{variable}{crime type}
\item{estimate}{crime rate per 100,000 inhabitants}
...
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/downloads}
}
\usage{
crime_by_msa_2024
}
\description{
A dataset with the rates at which various crime types occurred in
metropolitan statistical areas (MSAs) across the United States in 2024.
Modified from the U.S. Federal Bureau of Investigation's Crime in the United
States Annual Reports - 2024 CIUS Estimations, found at the given source
URL. Be careful in comparing estimates across years; see
\url{https://ucr.fbi.gov/ucr-statistics-their-proper-use} for additional
guidance on data use.
}
\keyword{datasets}
