% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_sheets_auth.R
\name{google_sheets_auth}
\alias{google_sheets_auth}
\title{Authorizes user account for Google Sheets data writing and retrieval in
\code{\link[=web_app]{web_app()}}}
\usage{
google_sheets_auth(email)
}
\arguments{
\item{email}{Supply the email address associated with your Google account.}
}
\description{
In order to collect and display data from app viewers in a
Shiny app created using \code{\link[=web_app]{web_app()}}, this function must first be executed so
that data can be written to and read from Google Sheets. This requires
a Google account. This function only needs to be run once, as your account
information will then be saved across R sessions.
}
