% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/place_geoids.R
\name{place_geoids}
\alias{place_geoids}
\title{Retrieves place GEOIDs for use in \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}}
\usage{
place_geoids(state = NULL, sf = FALSE, namelsad = TRUE)
}
\arguments{
\item{state}{Optionally, specify a vector of two-letter state abbreviations to filter
to. By default, all Census places (cities/towns/villages/census-designated
places) in U.S. states and territories will be returned by the function.}

\item{sf}{If \code{FALSE} (default value), function returns a data frame without
geospatial polygon geometry. If \code{TRUE}, function returns a shapefile with
geospatial polygon geometry.}

\item{namelsad}{If \code{TRUE} (default value), the returned dataset's \code{name}
variable will use a longer version of place names (e.g.
\code{"Hurtsboro town, Alabama"}) including each place's LSAD (legal/statistical
area description). If \code{FALSE}, the returned dataset's \code{name}
variable will use a shorter version of place names (e.g.
\code{"Hurtsboro, Alabama"}) excluding each place's LSAD.}
}
\description{
Using the tigris
\link[tigris:places]{places()} and \link[tigris:states]{states()}
functions, returns a dataset of places with their
corresponding states and GEOIDs, as designated by the U.S. Census Bureau for
the year 2023, for use in querying the \code{\link[=retrieve_aggregated]{retrieve_aggregated()}} function.
}
