% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_aggregated.R
\name{retrieve_aggregated}
\alias{retrieve_aggregated}
\alias{retrieve_aggregated.data.frame}
\alias{retrieve_aggregated.character}
\title{Retrieves and combines data, aggregated by geographic area, from various sources}
\usage{
retrieve_aggregated(
  vars,
  yrs = 1980:2024,
  geog,
  state = NULL,
  geoselect = NULL,
  vars_info = TRUE,
  sf = FALSE
)

\method{retrieve_aggregated}{data.frame}(
  vars,
  yrs = 1980:2024,
  geog,
  state = NULL,
  geoselect = NULL,
  vars_info = TRUE,
  sf = FALSE
)

\method{retrieve_aggregated}{character}(
  vars,
  yrs = 1980:2024,
  geog,
  state = NULL,
  geoselect = NULL,
  vars_info = TRUE,
  sf = FALSE
)
}
\arguments{
\item{vars}{Before supplying this argument, ensure you have read the
\code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} function description, and have run that function to
return the data frame of retrievable variables. The \code{vars} argument of this
\code{\link[=retrieve_aggregated]{retrieve_aggregated()}} function can be supplied in one of two ways:
\itemize{
\item As a vector of all variables you wish to retrieve. To retrieve all
variables within one or more categories as given in the
\code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} data frame, set \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}'s \code{vars}
argument equal to a call to the \code{\link[=categories_vars]{categories_vars()}} function, which will
return a vector of the desired variables.
\item As a data frame, specifically a version of the \code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}}
data frame filtered to only the rows corresponding to variables you wish to
retrieve.
}}

\item{yrs}{Supply a vector (in numeric format) of one or more years to filter
to. If no value is supplied, will
default to the largest possible range of years across which data is available
on at least one of the selected variables. The years for which data is
available for each variable are included in the \code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}}
dataset.}

\item{geog}{Specify \code{"county"}, \code{"cbsa"} (core based statistical
area), or \code{"place"} (cities/towns/villages/census-designated places). The
geographic area types for which data is available for each variable are
included in the \code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} data frame.}

\item{state}{Optionally (ONLY if \code{geog} is set to \code{"county"} or \code{"place"}),
supply a vector of state abbreviations to filter to. By default, all possible
values will be returned by the function.}

\item{geoselect}{Optionally, supply a vector (in character format) of county, CBSA, or place
(depending on whether \code{geog} is set to \code{"county"}, \code{"cbsa"}, or \code{"place"})
GEOIDs to filter to. By default, all possible values will be returned by the
function. To determine the corresponding GEOID for a particular county, CBSA,
or place, use \code{\link[=county_geoids]{county_geoids()}}, \code{\link[=cbsa_geoids]{cbsa_geoids()}}, or \code{\link[=place_geoids]{place_geoids()}},
respectively.}

\item{vars_info}{Specify whether to append columns containing
category, source, and type information for each variable in this function's
output data frame. Default value is \code{TRUE}. The appended information will
be the same as that contained in the \code{\link[=retrieve_aggregated_vars]{retrieve_aggregated_vars()}} data frame.}

\item{sf}{If \code{FALSE} (default value), function returns a data frame without
geospatial polygon geometry. If \code{TRUE}, function returns a shapefile with
geospatial polygon geometry.}
}
\description{
Retrieves data from sources such as the Census Bureau, Bureau of
Economic Analysis, and Environmental Protection Agency. User specifies
variables, timeframe, and geographic region of data to retrieve. Function
will return a dataset combining all variables requested.
}
