% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_vars.R
\name{retrieve_data_vars}
\alias{retrieve_data_vars}
\title{Returns table of variables retrievable through \code{\link[=retrieve_data]{retrieve_data()}}}
\usage{
retrieve_data_vars()
}
\description{
Returns a table with all variables retrievable through the
\code{\link[=retrieve_data]{retrieve_data()}} function. Variables are sorted
into ten categories: economy, housing & family, transportation &
walkability, health, education, crime, community spaces, sustainability,
governance, and demographics. For each variable, the table lists its
corresponding category as well as other information on the given variable
including whether an API key is required for retrieval, data source, data
type, and available years and geographic scales. If an API key is required
to retrieve a certain variable, run the indicated function for setting the
API key before attempting to retrieve the variable for the first time
using \code{\link[=retrieve_data]{retrieve_data()}}. Once set, the API key will be saved across R
sessions. Some variables are computed by cidar using multiple variables from
the data source. More information on how the variables were selected and
categorized is available at (add link to data documentation w/ explanation?).
}
