% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_point_polygon.R
\name{retrieve_point_polygon}
\alias{retrieve_point_polygon}
\alias{retrieve_point_polygon.character}
\alias{retrieve_point_polygon.list}
\title{Retrieves shapefile of point or polygon features from chosen source}
\usage{
retrieve_point_polygon(
  features,
  yrs = 1999:2024,
  geog = NULL,
  geoselect = NULL,
  dataset_info = TRUE
)

\method{retrieve_point_polygon}{character}(
  features,
  yrs = 1999:2024,
  geog = NULL,
  geoselect = NULL,
  dataset_info = TRUE
)

\method{retrieve_point_polygon}{list}(
  features,
  yrs = 1999:2024,
  geog = NULL,
  geoselect = NULL,
  dataset_info = TRUE
)
}
\arguments{
\item{features}{Before supplying this argument,
ensure you have read the \code{\link[=retrieve_point_polygon_features]{retrieve_point_polygon_features()}} function description, and have run that function to
return the data frame of retrievable features and their types. The \code{features}
argument for this \code{\link[=retrieve_point_polygon]{retrieve_point_polygon()}} function can be supplied in one of two ways:
\itemize{
\item Specify an atomic vector of feature sets to retrieve.
\item Specify a list of named vectors, where vector names are feature sets and
vector values are the types to include within each feature set. Set a vector
equal to \code{NULL} to include all possible types within the given feature set.
}}

\item{yrs}{If yrs is a numeric vector, this function will retrieve all
specified features for all years in the specified vector. If yrs is \code{NULL},
this function will retrieve all most recent years of selected features. If yrs is not provided,
retrieve all available years for all features associated with specific years.
The years for which data is available for each feature set are included in the
\code{\link[=retrieve_point_polygon_features]{retrieve_point_polygon_features()}} dataset. Please note that if \code{yrs} is
equal to \code{NA} in that dataset, the feature set can only be retrieved in
this \code{\link[=retrieve_point_polygon]{retrieve_point_polygon()}} function when \code{yrs} is set to \code{NULL}.}

\item{geog}{Optionally, specify a geographic area to filter to. Can be
\code{"state"}, \code{"county"}, \code{"cbsa"} (core based statistical
area), \code{"place"}, or \code{"zip code"}. If no value is provided, a nationwide
dataset will be returned. Please see \code{\link[=retrieve_point_polygon_features]{retrieve_point_polygon_features()}} to
determine whether a particular feature set is able to be filtered to a particular
geographic area, and only specify a geographic area that all selected feature
sets can be filtered to.}

\item{geoselect}{If a value is supplied for \code{geog}, supply a vector (in character format) of two-letter state abbreviations; county,
CBSA, or place GEOIDS; or zip codes (depending on whether \code{geog} is set to \code{"state"},
\code{"county"}, \code{"cbsa"} or \code{"place"}, or \code{"zip code"}) to filter to. By default, all possible
values will be returned by the
function. To determine the corresponding GEOID for a particular county, CBSA,
or place, use \code{\link[=county_geoids]{county_geoids()}}, \code{\link[=cbsa_geoids]{cbsa_geoids()}}, or \code{\link[=place_geoids]{place_geoids()}},
respectively. Note that for polygon features, all features intersecting the
specified geographic area will be returned, not just those fully inside of it.}

\item{dataset_info}{Specify whether to append columns containing
source information for the dataset in this function's
output. Default value is \code{TRUE}. The appended information will
be the same as that contained in the \code{\link[=retrieve_point_polygon_features]{retrieve_point_polygon_features()}} data frame.}
}
\description{
Retrieves shapefile of point or polygon features of user-specified data category,
from data sources such as ArcGIS Online and the Institute of Museum and Library Services.
}
