% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time.R
\name{travel_time}
\alias{travel_time}
\title{Assesses travel times to various destination types}
\usage{
travel_time(destination, geog, geoselect, mins, mode)
}
\arguments{
\item{destination}{Destination type to which to calculate travel time.
Available destination types include: \code{"schools"}, \code{"public schools"}, and
\code{"private schools"}.}

\item{geog}{Type of area in which to compute travel time radii from selected
destination type. Specify \code{"county"}, \code{"cbsa"}, or \code{"zip code"}.}

\item{geoselect}{Supply a specific county GEOID (if \code{geog} is set to
\code{"county"}), CBSA GEOID (if \code{geog} is set to \code{"cbsa"}), or zip code (if
\code{geog} is set to \code{"zip code"}) within which to compute travel time radii from
selected destination type. To determine the corresponding GEOID for a
particular county or CBSA, use \code{\link[=county_geoids]{county_geoids()}} or \code{\link[=cbsa_geoids]{cbsa_geoids()}},
respectively.}

\item{mins}{Number of minutes traveling.}

\item{mode}{Mode of travel. Available modes include: \code{"driving"}, \code{"cycling"}
, and \code{"walking"}.}
}
\description{
Provides travel times by foot, bicycle, or car to destination
types, including schools, (will add more), in a specified area. Utilizes the
TravelTime API, called through the traveltimeR package. Requires your
TravelTime application ID and application key to already be saved, which can
be done using \code{\link[=travel_time_key]{travel_time_key()}}.
}
