% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_time_key.R
\name{travel_time_key}
\alias{travel_time_key}
\title{Establishes a TravelTime API ID and key for \code{\link[=travel_time]{travel_time()}}}
\usage{
travel_time_key(id, key)
}
\arguments{
\item{id}{Supply your TravelTime application ID.}

\item{key}{Supply your TravelTime application key.}
}
\description{
Saves your TravelTime API ID and key so you won't have to
reenter it, even between R sessions. If you haven't yet registered for the
TravelTime API, follow the instructions at https://account.traveltime.com.
}
