% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbsa_geoids.R
\name{cbsa_geoids}
\alias{cbsa_geoids}
\title{Retrieves CBSA GEOIDs for use in \code{\link[=retrieve_data]{retrieve_data()}} and \code{\link[=travel_time]{travel_time()}}}
\usage{
cbsa_geoids()
}
\description{
Returns a table of U.S. CBSAs (Core Based Statistical Areas)
with their corresponding GEOIDs, for use in querying the \code{\link[=retrieve_data]{retrieve_data()}}
and \code{\link[=travel_time]{travel_time()}} functions. If you have not already set a U.S. Census
Bureau API key using \code{\link[=census_key]{census_key()}}, do so before running this function.
}
