% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_key.R
\name{census_key}
\alias{census_key}
\title{Establishes a U.S. Census Bureau API key for \code{\link[=retrieve_data]{retrieve_data()}}}
\usage{
census_key(key)
}
\arguments{
\item{key}{Supply your Census Bureau API key.}
}
\description{
If you haven't yet registered for a Census Bureau API key,
follow the instructions at https://api.census.gov/data/key_signup.html.
}
