% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_geoids.R
\name{county_geoids}
\alias{county_geoids}
\title{Retrieves county GEOIDs for use in \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}}
\usage{
county_geoids(state = NULL, sf = FALSE, namelsad = TRUE)
}
\arguments{
\item{state}{Optionally, specify a vector of two-letter state abbreviations to filter
to. By default, all Census-designated counties in U.S. states and territories
will be returned by the function.}

\item{sf}{If \code{FALSE} (default value), function returns a data frame without
geospatial polygon geometry. If \code{TRUE}, function returns a shapefile with
geospatial polygon geometry.
@param namelsad If \code{TRUE} (default value), the returned dataset's \code{name}
variable will use a longer version of county names (e.g.
\code{"Autauga County, Alabama"}) including each county's LSAD (legal/
statistical area description). If \code{FALSE}, the returned dataset's \code{name}
variable will use a shorter version of county names (e.g.
\code{"Autauga, Alabama"}) excluding each county's LSAD.}
}
\description{
Using the tigris
\link[tigris:counties]{counties()} and \link[tigris:states]{states()}
functions, returns a dataset of counties with their
corresponding states and GEOIDs, as designated by the U.S. Census Bureau for
the year 2023, for use in querying the \code{\link[=retrieve_aggregated]{retrieve_aggregated()}} function.
}
