% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_geoids.R
\name{county_geoids}
\alias{county_geoids}
\title{Retrieves county GEOIDs for use in \code{\link[=retrieve_data]{retrieve_data()}} and \code{\link[=travel_time]{travel_time()}}}
\usage{
county_geoids(state = NULL)
}
\arguments{
\item{state}{Optionally, specify a vector of state abbreviations to filter
to. By default, counties in all U.S. states will be returned by the function.}
}
\description{
Returns a table of U.S. counties with their corresponding
GEOIDs, for use in querying the \code{\link[=retrieve_data]{retrieve_data()}} and \code{\link[=travel_time]{travel_time()}}
functions. If you have not already set a U.S. Census Bureau API key using
\code{\link[=census_key]{census_key()}}, do so before running this function.
}
