% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crime_by_msa_2023}
\alias{crime_by_msa_2023}
\title{Crimes by Metropolitan Statistical Area in the United States in 2023}
\format{
A data frame with 2988 rows and 4 variables:
\describe{
\item{msa}{metropolitan statistical area}
\item{variable}{crime type}
\item{estimate}{crime rate}
\item{year}{year (always 2023)}
...
}
}
\source{
\url{https://cde.ucr.cjis.gov/LATEST/webapp/#/pages/downloads}
}
\usage{
crime_by_msa_2023
}
\description{
A dataset with the rates at which various crime types occurred in metropolitan
statistical areas (MSAs) across the United States in 2023. Modified from the
given URL.
}
\keyword{datasets}
