% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_aggregated_vars.R
\name{retrieve_aggregated_vars}
\alias{retrieve_aggregated_vars}
\title{Returns table of variables retrievable through \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}}
\usage{
retrieve_aggregated_vars()
}
\description{
Returns a table with all variables retrievable through the
\code{\link[=retrieve_aggregated]{retrieve_aggregated()}} function. Variables are sorted
into ten categories: economy, housing & family, transportation &
walkability, health, education, crime, community spaces, sustainability,
governance, and demographics. For each variable, the table lists its
corresponding category as well as other information on the given variable
including whether an API key is required for retrieval, data source, data
type, and available years and geographic scales. If an API key is required
to retrieve a certain variable, run the indicated function for setting the
API key before attempting to retrieve the variable for the first time
using \code{\link[=retrieve_aggregated]{retrieve_aggregated()}}. Once set, the API key will be saved across R
sessions. Some variables are computed by cidar using multiple variables from
the data source. More information on how the variables were selected and
categorized is available at (add link to data documentation w/ explanation?).
}
\details{
Details on included datasets:
\itemize{
\item Census Bureau API - American Community Survey 5-year estimates: These estimates should only be compared with each other across non-overlapping 5-year periods. See \url{https://www.census.gov/programs-surveys/acs/guidance/comparing-acs-data.html} for more guidance on data use.
\item U.S. Environmental Protection Agency: See \url{https://aqs.epa.gov/aqsweb/airdata/download_files.html} for more information on data files.
\item U.S. Federal Bureau of Investigation - Crime in the United States estimations: Be careful in comparing estimates across years; see \url{https://ucr.fbi.gov/ucr-statistics-their-proper-use} for additional guidance on data use. Note that the definition of rape was changed between 2012 and 2013, see \url{https://ucr.fbi.gov/crime-in-the-u.s/2013/crime-in-the-u.s.-2013/rape-addendum/rape_addendum_final} for more details.
}
}
