% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data.R
\name{retrieve_data}
\alias{retrieve_data}
\title{Retrieves and combines data from various sources}
\usage{
retrieve_data(vars, yrs = 1980:2024, geog, state = NULL, geoselect = NULL)
}
\arguments{
\item{vars}{Supply a vector of variables to request. Available variables include: unemployment, high-skill employment, median family income, poverty, old-age dependency ratio, housing vacancy, GDP per capita, median AQI. Some variables cannot be retrieved without first registering for an API key for the corresponding data source and supplying the API key in R (see notes below).
\itemize{
\item Variables requiring a U.S. Census Bureau API key: unemployment, high-skill employment, median family income, poverty, old-age dependency ratio, housing vacancy. To retrieve these variables, first establish a Census Bureau API key by following the instructions for \code{\link[=census_key]{census_key()}}.
\item Variables requiring a U.S. Bureau of Economic Analysis API key: GDP per capita. To retrieve these variables, first establish a Bureau of Economic Analysis API key by following the instructions for \code{\link[=bea_key]{bea_key()}}.
}}

\item{yrs}{Supply a year or range of years. If no value is supplied, will default to the largest possible range of years across which data is available on at least one of the selected variables.}

\item{geog}{Specify either "county" or "cbsa".}

\item{state}{Optionally (ONLY if geog is set to "county"), supply a vector of state abbreviations for filter to. By default, all possible values will be returned by the function.}

\item{geoselect}{Optionally, supply a vector of county or CBSA (depending on whether geog is set to "county" or "cbsa") GEOIDs to filter to. By default, all possible values will be returned by the function.}
}
\description{
Retrieves data from sources such as the Census Bureau, Bureau of Economic Analysis, and Environmental Protection Agency. User specifies variables, timeframe, and geographic region of data to retrieve. Function will return a dataframe combining all variables requested.
}
