% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_point_polygon.R
\name{retrieve_point_polygon}
\alias{retrieve_point_polygon}
\title{Retrieves shapefile of point or polygon features from chosen source}
\usage{
retrieve_point_polygon(
  features,
  geog = NULL,
  geoselect = NULL,
  dataset_info = TRUE
)
}
\arguments{
\item{features}{Specify a vector of feature sets to retrieve. Before supplying this argument,
ensure you have read the (link to retrieve_point_polygon_features) function description, and have run that function to
return the data frame of retrievable features.}

\item{geog}{Optionally, specify a geographic area to filter to. Can be
\code{"state"}, \code{"county"}, \code{"cbsa"} (core based statistical
area), \code{"place"}, or \code{"zip code"}. If no value is provided, a nationwide
dataset will be returned. Please see (link to retrieve_point_polygon_features) to
determine whether a particular feature set is able to be filtered to a particular
geographic area, and only specify a geographic area that all selected feature
sets can be filtered to.}

\item{geoselect}{If a value is supplied for \code{geog}, supply a vector of state abbreviations, county,
CBSA, or place GEOIDS, or zip codes (depending on whether \code{geog} is set to \code{"state"},
\code{"county"}, \code{"cbsa"} or \code{"place"}, or \code{"zip code"}) to filter to. By default, all possible
values will be returned by the
function. To determine the corresponding GEOID for a particular county, CBSA,
or place, use \code{\link[=county_geoids]{county_geoids()}}, \code{\link[=cbsa_geoids]{cbsa_geoids()}}, or \code{\link[=place_geoids]{place_geoids()}},
respectively. Note that for polygon features, all features intersecting the
specified geographic area will be returned, not just those fully inside of it.}

\item{dataset_info}{Specify whether to append columns containing
category, source, and type information for the dataset in this function's
output. Default value is \code{TRUE}. The appended information will
be the same as that contained in the (link to retrieve_point_polygon_features) data frame.}
}
\description{
Retrieves shapefile of point or polygon features of user-specified data category. (discuss data sources)
}
