% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_point_polygon_features.R
\name{retrieve_point_polygon_features}
\alias{retrieve_point_polygon_features}
\title{Returns table of geospatial feature sets retrievable through \code{\link[=retrieve_point_polygon]{retrieve_point_polygon()}}}
\usage{
retrieve_point_polygon_features()
}
\description{
Returns a table with all point and polygon feature sets
retrievable through the \code{\link[=retrieve_point_polygon]{retrieve_point_polygon()}} function. For each set of
features, the table lists the geometry type (point or polygon), whether
an API key is required for data retrieval, what the data source is, what
subtypes the feature set is categorized into ("types"), what year the feature set is
from (this will either be the most recent year available, or say that the
dataset is updated regularly), and what geography types
the feature set can be filtered by. If an API key is required
to retrieve a certain variable, run the indicated function for setting the
API key before attempting to retrieve the variable for the first time
using \code{\link[=retrieve_point_polygon]{retrieve_point_polygon()}}. Once set, the API key will be saved across R
sessions.
}
