% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_app.R
\name{web_app}
\alias{web_app}
\title{Creates Shiny app to assist in sharing and collecting community-focused data}
\usage{
web_app(
  project_name,
  title,
  response_questions,
  shapefile_vars,
  shapefile_features
)
}
\arguments{
\item{project_name}{Name of project you are running this function for. Will
use the same Google Sheets document as any other time you call this function.
The first time you specify a certain project name, this function will
create a corresponding Google Sheets document.}

\item{title}{Title to display at the top of the web app.}

\item{response_questions}{Vector of prompts for web app viewers on the type
of data you are looking for from them. Each prompt will appear as a different
question, with data to be included on a different sheet in the Google Sheets
document.}

\item{shapefile_vars}{Shapefile of data, specifically with numeric variables
that can be color-coded, to display on map alongside a slider to select a variable
to display. Shapefile must contain
a \code{"year"} variable (the web app will include a selector slider for
years), a \code{"variable"} variable (the web app will include a dropdown for
viewers to select a variable for display, e.g. unemployment, median air quality), an
\code{"estimate"} variable (for the value or
estimate associated with each value in the \code{"variable"} column), and a \code{"name"} variable containing
the names of the geographic areas (e.g. Orange County), which will be
displayed alongside estimates upon mouse-over of the polygons. The
shapefile's geometry must be polygons.}

\item{shapefile_features}{Shapefile of points and/or polygons, without a corresponding
numeric variable. All points and polygons in this
shapefile will display simultaneously by default. Shapefile must contain
a \code{"features"} variable (for the name of the feature category, e.g. park, school),
a \code{"name"} variable (for the name of each individual feature, e.g. Central Park),
and a \code{"type"} variable (delineating different feature types within each feature
category, e.g. private school, public school). All observations corresponding
to a given value of the \code{"features"} variable must be either points or
polygons (i.e. no mixing of points and polygons within a feature category).}
}
\description{
Creates web app using the \link[shiny:shiny-package]{shiny} package,
with options for visualizing your existing data and analysis for community
sharing, as well as for visualizing and collecting data from app viewers
(e.g. community members). If you have not yet done so, you must run
\code{\link[=google_sheets_auth]{google_sheets_auth()}}, following the input instructions, before being able
to visualize and collect data from viewers through this function, which
will draw from the \link[googlesheets4:googlesheets4-package]{googlesheets4}
package functionality to do so.'
}
