% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/web_app.R
\name{web_app}
\alias{web_app}
\title{Creates Shiny app to assist in sharing and collecting community-focused data}
\usage{
web_app(response_question, sheet_id, sheet_name)
}
\arguments{
\item{response_question}{Prompt for web app viewers on the type of data you
are looking for from them}

\item{sheet_id}{ID of the Google Sheet to use (WILL ADD FURTHER INSTRUCTIONS)}

\item{sheet_name}{Name of the sheet to use within Google Sheets document}
}
\description{
Creates web app using the R Shiny package, with options for
visualizing your existing data and analysis for community sharing, as well
as for visualizing and collecting data from app viewers (e.g. community
members). If you have not yet done so, you must run \code{\link[=google_sheets_auth]{google_sheets_auth()}},
following the input instructions, before being able to visualize and collect
data from viewers through this function.
}
